# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Staffing Platform API Doc
  description: |
    On-Demand Healthcare Staffing & Visit Management Portal — REST API.

    ## Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtain tokens via `POST /auth/login`. Access tokens expire in **15 minutes**;
    use `POST /auth/refresh` with a valid refresh token to rotate.

    ## Roles
    | Role | Description |
    |------|-------------|
    | `SUPER_ADMIN` | Full platform control |
    | `RECRUITER` | Read-only admin access |
    | `FACILITY_ADMIN` | Manages one facility tenant |
    | `TEAM_MEMBER` | Scoped facility operator |
    | `NURSE` | Field professional (mobile) |

    ## File Uploads
    All uploads are stored in **DigitalOcean Spaces** (private).
    Use `POST /storage/signed-upload-url` for direct client-to-Spaces uploads,
    or `POST /storage/upload/{folder}` for server-proxied multipart uploads.
    Download links are short-lived signed URLs — never public.

    ## Pagination
    List endpoints accept `page` and `limit` query params and return:
    ```json
    { "pagination": { "page": 1, "limit": 20, "total": 100, "totalPages": 5 } }
    ```

    ## EVV Geofencing
    Check-in/out enforces a **200-metre geofence** around the case address (Haversine).
    Breaches create a `FLAGGED` visit requiring admin override approval.
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Parameters ──────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but insufficient role/scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    CredentialType:
      type: string
      enum: [STATE_LICENSE, CPR_CERTIFICATION, TB_TEST, BACKGROUND_CHECK,
             GOVERNMENT_ID, OIG_CHECK, SAM_CHECK, IMMUNIZATION,
             WORK_AUTHORIZATION, CUSTOM]
    CredentialStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED]
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:    { type: object }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 'Resource not found' }
...
    PaginationMeta:
      type: object
      properties:
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }

    # ── Auth ───────────────────────────────
...
    CredentialSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        type:      { $ref: '#/components/schemas/CredentialType' }
        status:    { $ref: '#/components/schemas/CredentialStatus' }
        expiresAt: { type: string, format: date-time, nullable: true }
    CredentialDetail:
      allOf:
        - $ref: '#/components/schemas/CredentialSummary'
        - type: object
          properties:
            customLabel:     { type: string, nullable: true }
            fileUrl:         { type: string }
            downloadUrl:     { type: string, description: 'Short-lived signed URL' }
            issuedAt:        { type: string, format: date-time, nullable: true }
            rejectionReason: { type: string, nullable: true }
            reviewedAt:      { type: string, format: date-time, nullable: true }
            createdAt:       { type: string, format: date-time }

    # ── Facilities ─────────────────────────
...
security:
  - BearerAuth: []

# docs/paths-part1.yaml  (auth · users · credentials · facilities)
paths:
...
  /credentials:
    post:
      tags: [Credentials]
      summary: Upload a credential document
      description: |
        **Roles:** NURSE  
        Accepts `multipart/form-data`. File is stored privately in DigitalOcean Spaces.
        Status defaults to `PENDING` until reviewed by an admin.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                  description: 'PDF, JPG, PNG or WEBP, max 15 MB'
                type:        { $ref: '#/components/schemas/CredentialType' }
                customLabel: { type: string, description: 'Required when type=CUSTOM' }
                issuedAt:    { type: string, format: date }
                expiresAt:   { type: string, format: date }
      responses:
        '201':
          description: Credential uploaded and pending review
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CredentialDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413':
          description: File too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    get:
      tags: [Credentials]
      summary: List all credentials (admin/recruiter)
      description: '**Roles:** SUPER_ADMIN, RECRUITER'
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/CredentialStatus' }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/CredentialType' }
      responses:
        '200':
          description: Paginated credential list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CredentialDetail' }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }
...
  /credentials/admin/expiry-alerts:
    get:
      tags: [Credentials]
      summary: Credentials expiring within N days
      description: '**Roles:** SUPER_ADMIN, RECRUITER'
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
          description: Look-ahead window in days
      responses:
        '200':
          description: Expiring credentials
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CredentialDetail' }
  /credentials/{id}:
    get:
      tags: [Credentials]
      summary: Get a credential with signed download URL
      description: Nurses can only fetch their own credentials.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Credential detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CredentialDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Credentials]
      summary: Delete an unreviewed credential
      description: '**Roles:** NURSE — cannot delete approved credentials.'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Credential deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Cannot delete an approved credential
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /credentials/{id}/approve:
    patch:
      tags: [Credentials]
      summary: Approve a credential
      description: '**Roles:** SUPER_ADMIN, RECRUITER. Triggers push + email notification to nurse.'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Credential approved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CredentialDetail' }
  /credentials/{id}/reject:
    patch:
      tags: [Credentials]
      summary: Reject a credential with a reason
      description: '**Roles:** SUPER_ADMIN, RECRUITER'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Credential rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }

  # ═══════════════════════════════════════════
  # FACILITIES
  # ═══════════════════════════════════════════
...